+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/management/ui/template/UITemplateInitializer.java
@@ -65,6 +65,7 @@ public class UITemplateInitializer {
         Layer.AWS_S3.name(),
         Layer.AWS_DYNAMODB.name(),
         Layer.AWS_GATEWAY.name(),
+        Layer.REDIS.name(),
         "custom"
     };
     private final UITemplateManagementService uiTemplateManagementService;